+++ CHANGELOG.md
@@ -5,12 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- Remove ```Router::view```
 
 ## [0.1.2] - 2021-09-15
 ### Added
 - Add class to render view ```View::class```
 ### Fixed 
--  ```Router::class``` view cant render using child controller class
+-  ```Router::view``` cant render using child controller class
 
 ## [0.1.1] - 2021-09-14
 ### Added